requirements-pipeline.txt
pandas
numpy
requests
python-dotenv
tenacity
PyYAML
rapidfuzz
reportlab
qrcode
pillow
tqdm
